backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.7
Pillow==10.0.1
python-decouple==3.8
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
cryptography==41.0.7
h11==0.16.0
huggingface-hub==0.17.3
transformers==4.35.2
torch==2.1.1
torchvision==0.16.1
pytesseract==0.3.10
opencv-python==4.8.1.78
pandas==2.1.3
numpy==1.25.2
groq==0.4.1
python-dotenv==1.0.0
requests==2.31.0
httpx==0.27.2
pdfplumber==0.11.4